refactor: Layout 세팅 코드를 확장 파일로 분리
