+++ iOS/Projects/Features/Profile/Sources/Presentation/ProfileScene/ProfileLayouts/SectionItem.swift
@@ -1,0 +1,19 @@
+//
+//  SectionItem.swift
+//  ProfileFeature
+//
+//  Created by 홍승현 on 12/5/23.
+//  Copyright © 2023 kr.codesquad.boostcamp8. All rights reserved.
+//
+
+import Foundation
+
+// MARK: - ProfileSection
+
+enum ProfileSection: Int {
+  case header
+  case main
+  case emptyState
+}
+
+typealias ProfileItem = String
